bump mkuimage to use -skip-ldd flag
With the recent update in github.com/u-root/mkuimage, there is a
new flag available to control the resolution of ldd dependencies
when including binary files with the -files flag.

The flag will be registered automatically by the 'RegisterFlags()'
method of 'mkuimage.Flags' in u-root's 'main()', there is no need
to also define the flag in here in u-root.

The README was synced with the respective part of the README in
github.com/u-root/mkuimage.

Signed-off-by: Jens Topp <jens.topp@9elements.com>