add test  using github.com/u-root/cpu{vm,client}
This is a simple test, which needs no shell,
demonstrating how to use github.com/u-root/cpu vm and client packages.

The vm package has four Linux kernel/initramfs pairs embedded in it,
for amd64, arm, arm64, and riscv64; Plan 9 is planned.

The test uses vm.Uroot to build a complete u-root image and merge
it into the provided initramfs.

The vm package will start the VM (see the test) with cpud running.

The test then uses the client package to run commands in the guest,
without needing a shell script.

The cpu client returns 0 or non-zero status, for command success of failure.

Note that, also, files can be read or written from outside the vm, with no
special effort. This is shown in the first part of the tests, via dd.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>