ip_linux test using the cpu command
This test uses cpu to implement the ip test.

It should be equivalent to the shell-script-based command.

It is a copy of the one in github.com/u-root/cpu/vm,
which is used to test the vm package.

This PR includes changes for two other tests to
catch up to u-root/cpu/vm tip.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>