set failok to false
Turns out the kernel in ghcr.io is modular (!!)
and it seems the shell script test is working
by an accident I don't understand, possibly the kernel
shelling out to modprobe?

I'd rather it be a non-modular kernel, and this test
should never have failures. So back to failing until
I can dity up the ghcr.io kernel.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>