mv extra/tsort to cmds/tsort
In a talk at #OSFC25, a speaker mentioned that the existince
of extra/ was a bit confusing.

They are right. There's only ever been one command in there,
it's small, and there is no reason not to have it in core.

The existence of extra/ seems to have resulted from a misunderstanding.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>