Only reserve MMCONFIG on x86 platforms
Reserving the MMCONFIG is required on x86 platforms to ensure that PCI
devices work properly in the OS. In ARM, UPL uses device trees to
discover PCI devices, which can cause conflicts if the MMCONFIG regions
are reserved.

Tested: Booted UPL on ARM platform and verified boot. Booted x86
platform and verified regions are reserved.

Signed-off-by: Andrew Sun <andrewsun@google.com>