Fix an ordering issue in menu
Under some environments the fact that the menu actually switches form
stdio to /dev/tty in the middle of processing actually matters.

This is done because in some environments the stdio devices don't
really handle input timeouts well, but the downside is that they
can be two very different data paths. When the file descriptor used
for the prompt is put into raw mode the buffering in the other file
descriptor is not notified so any data not flushed out remains
unflushed. The end result is that the prompt, sent out last can show up
on the screen at almost any place from the top to the bottom.

Since we can't just use the stdout device (if we trust the comments that
say that we can't) then the only other option is to force a flush on all
the data sent before the prompt (and throw in a tiny delay for safety).

Fixes: #3426
Signed-off-by: jelischer <58259690+jelischer@users.noreply.github.com>