Add the builtin package, with a forth example
The builtin package is an proposal to
implementing command builtins. The core idea is
that a package can be used as a builtin, without
using the gobusybox.

It is only for special circumstances where use of the
gobusybox is not practical.

Gobusybox remains the preferred approach, since it
can incorporate arbitrary Go programs outside u-root.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>